requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
facenet-pytorch==2.2.9
fastapi==0.108.0
filelock==3.9.0
fsspec==2023.4.0
h11==0.14.0
idna==3.6
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.2.1
networkx==3.0rc1
numpy==1.26.3
opencv-python==4.9.0.80
pillow==10.2.0
psycopg==3.1.17
pydantic==2.5.3
pydantic_core==2.14.6
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
starlette==0.32.0.post1
sympy==1.11.1
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.25.0
